documentation and move verify to test
